+++ src/commonMain/kotlin/com/algolia/search/model/response/ResponseHitWithPosition.kt
@@ -1,0 +1,11 @@
+package com.algolia.search.model.response
+
+import kotlinx.serialization.Serializable
+
+
+@Serializable
+public data class ResponseHitWithPosition(
+    val hit: ResponseSearch.Hit,
+    val position: Int,
+    val page: Int
+)