+++ src/commonTest/kotlin/serialize/TestKeys.kt
@@ -353,5 +353,7 @@ internal class TestKeys {
         KeyVillage shouldEqual "village"
         KeySimilarQuery shouldEqual "similarQuery"
         KeyEnableABTest shouldEqual "enableABTest"
+        KeyAlternatives shouldEqual "alternatives"
+        KeyIndexLanguages shouldEqual "indexLanguages"
     }
 }