+++ src/commonTest/kotlin/serialize/rule/TestAlternatives.kt
@@ -1,0 +1,14 @@
+package serialize.rule
+
+import com.algolia.search.model.rule.Alternatives
+import kotlinx.serialization.json.JsonLiteral
+import serialize.TestSerializer
+
+
+internal class TestAlternatives : TestSerializer<Alternatives>(Alternatives) {
+
+    override val items = listOf(
+        Alternatives.True to JsonLiteral(true),
+        Alternatives.False to JsonLiteral(false)
+    )
+}