+++ src/commonTest/kotlin/serialize/rule/TestCondition.kt
@@ -1,13 +1,11 @@
 package serialize.rule
 
 import attributeA
+import com.algolia.search.model.rule.Alternatives
 import com.algolia.search.model.rule.Anchoring
 import com.algolia.search.model.rule.Condition
 import com.algolia.search.model.rule.Pattern
-import com.algolia.search.serialize.JsonNoDefaults
-import com.algolia.search.serialize.KeyAnchoring
-import com.algolia.search.serialize.KeyIs
-import com.algolia.search.serialize.KeyPattern
+import com.algolia.search.serialize.*
 import kotlinx.serialization.json.json
 import serialize.TestSerializer
 import unknown
@@ -18,17 +16,21 @@ internal class TestCondition : TestSerializer<Condition>(Condition.serializer(),
     override val items = listOf(
         Condition(
             anchoring = Anchoring.Is,
-            pattern = Pattern.Facet(attributeA)
+            pattern = Pattern.Facet(attributeA),
+            alternative = Alternatives.True
         ) to json {
             KeyAnchoring to KeyIs
             KeyPattern to "{facet:$attributeA}"
+            KeyAlternatives to true
         },
         Condition(
             anchoring = Anchoring.Is,
-            pattern = Pattern.Literal(unknown)
+            pattern = Pattern.Literal(unknown),
+            alternative = Alternatives.False
         ) to json {
             KeyAnchoring to KeyIs
             KeyPattern to unknown
+            KeyAlternatives to false
         }
     )
 }