+++ src/commonTest/kotlin/serialize/settings/TestSettings.kt
@@ -84,7 +84,8 @@ internal class TestSettings : TestSerializer<Settings>(Settings.serializer()) {
             maxFacetHits = int,
             version = int,
             advancedSyntaxFeatures = listOf(AdvancedSyntaxFeatures.ExcludeWords, AdvancedSyntaxFeatures.ExactPhrase),
-            userData = json { unknown to unknown }
+            userData = json { unknown to unknown },
+            indexLanguages = listOf(Language.Japanese)
         ) to json {
             // Attributes
             KeySearchableAttributes to attributesJson
@@ -154,6 +155,7 @@ internal class TestSettings : TestSerializer<Settings>(Settings.serializer()) {
             KeyVersion to int
             KeyUserData to json { unknown to unknown }
             KeyPrimary to JsonNull
+            KeyIndexLanguages to jsonArray { +Language.Japanese.raw }
         }
     )