+++ src/commonTest/kotlin/suite/TestSuiteSearch.kt
@@ -3,6 +3,8 @@ package suite
 import clientAdmin1
 import clientSearch
 import com.algolia.search.helper.toAttribute
+import com.algolia.search.model.ObjectID
+import com.algolia.search.model.response.ResponseSearch
 import com.algolia.search.model.search.Query
 import com.algolia.search.model.settings.AttributeForFaceting
 import com.algolia.search.model.settings.Settings
@@ -11,6 +13,7 @@ import com.algolia.search.model.task.TaskStatus
 import kotlinx.serialization.json.JsonObjectSerializer
 import kotlinx.serialization.list
 import runBlocking
+import shouldBeNull
 import shouldBeTrue
 import shouldContain
 import shouldEqual
@@ -46,13 +49,33 @@ internal class TestSuiteSearch {
             index.apply {
                 tasks += setSettings(settings)
                 tasks += saveObjects(objects)
-
                 tasks.wait().all { it is TaskStatus.Published }.shouldBeTrue()
-                search(Query("algolia")).nbHits shouldEqual 2
-                val hits = search(Query("elon", clickAnalytics = true)).hits
 
-                hits.shouldNotBeNull()
-                hits.shouldNotBeEmpty()
+                search(Query("algolia")).apply {
+                    nbHits shouldEqual 2
+                    getObjectIDPosition(ObjectID("nicolas-dessaigne")) shouldEqual 0
+                    getObjectIDPosition(ObjectID("julien-lemoine")) shouldEqual 1
+                    getObjectIDPosition(ObjectID("unknown")) shouldEqual -1
+                }
+
+                findFirstObject({ false }, Query(""), false).shouldBeNull()
+                findFirstObject({ true }, Query(""), false)!!.apply {
+                    page shouldEqual 0
+                    position shouldEqual 0
+                }
+                val predicate = { hit: ResponseSearch.Hit -> hit.json.getPrimitive("company").content == "Apple" }
+
+                findFirstObject(predicate, Query("Algolia"), false).shouldBeNull()
+                findFirstObject(predicate, Query(hitsPerPage = 5), true).shouldBeNull()
+                findFirstObject(predicate, Query(hitsPerPage = 5), false)!!.apply {
+                    position shouldEqual 0
+                    page shouldEqual 2
+                }
+
+               search(Query("elon", clickAnalytics = true)).apply {
+                   hits.shouldNotBeNull()
+                   hits.shouldNotBeEmpty()
+               }
             }
             search.apply {
                 search(Query(facets = allFacets, filters = "company:tesla")).nbHits shouldEqual 1